+++ RELEASE_NOTES.md
@@ -5,8 +5,11 @@
 - **Tables Overview**:
   - Added optional column "Tutorials" also to the regular table overview. This way, users can easier see if there will be a tutorial available in the pause menu.
   - Improved ROM column: The tooltip of a ROM displays additional alias information now. In addition to that, you can use the table search field to filter games by their ROM name or alias.
+  - Improved VPin MAME section: In addition to the alias mapping, the section shows also the tables which share the same ROM.
 - **Pause Menu**
   - Added pause menu support for stand-alone mode of VPX.
+- **Monitoring View / Screenshots**
+  - Added a summary screenshot to the screenshot bundle which contains all screens as one image.
 
 ## Bugfixes